==> code/01A_Gathering_Reddit_Data.ipynb <==
# Reddit & Quibi: Web API and NLP
## Part 1-A: Gathering Data from Reddit with Pushshift API

Quibi is a new mobile-only streaming platform that launched April 2020. All of the shows are serial and have episodes under 10 minutes. They have lots of big names doing shows such as Chrissy Tiegen and Sophie Turner. Quibi stands for "quick bites" as the content is meant to be consumed during "in between" times of your day. Since Quibi is brand new, I want to analyze how well their content compares to that of other popular media: **videos, television, and podcasts**.
- [Videos](https://www.reddit.com/r/videos/): The length of the content is most similar to videos found on services like YouTube and Vimeo
- [Television](https://www.reddit.com/r/cordcutters/): Given the start power and financial backing of each project, they are definitely going for the narrative and production quality of television shows. I'm specifically choosing the cordcutters reddit because as of 2019, [more people pay for streaming than for cable](https://fortune.com/2019/03/19/cord-cutting-record-netflix-deloitte/). This is especially true for Millenials who are the target demographic for Quibi.
- [Podcasts](https://www.reddit.com/r/podcasts/): As mentioned earlier, the content is meant to be consumed while you're waiting or in between tasks. Podcasts are often consumed in a similar manner. 

To see what people are saying about each of these area, I'm going to use the Pushshift API to gather 30,000 of the most recent posts for each of the subreddits above.

**Problem Statement**: How can we best segment the Quibi slate to reach audiences that enjoy videos, television, and podcasts?

import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import time

Since the Pushshift API has a limit of 1000 posts per request, the function below will take in a specific subreddit and number of desired posts, then gather until the number is reached. It will also print out periodic updates of how many have been gathered and the date from the oldest post.

#Function to get posts from reddit
def get_posts(subreddit, num):
    #Setting the base url and first "before" time
    base_url = 'https://api.pushshift.io/reddit/submission/search'
    bef_time = t
    
    #list to hold the dataframes to concat
    to_concat = []
    
    #While loop that keeps gathering until the number of desired posts is reached
    while len(to_concat) < (num / 1000):
        params = {
            'subreddit' : subreddit,
            'size' : 1000,
            'before' : bef_time,
            'lang' : True,
            'author': '![deleted]'
                }
        get = requests.get(base_url, params)
        data = get.json()['data']
        df = pd.DataFrame(data)
        bef_time = df['created_utc'].min()
        to_concat.append(df)
        
        #If statement to print out updates every 5000 posts including the time of the earliest post
        if len(to_concat) % 5 == 0:
            #Converting the epoch time into a more readable, datetime format
            print_time = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(bef_time))
            print(f'{len(to_concat)*1000} posts have been gathered, oldest post is from {print_time}')
    
    #Once the while loop is done, concat the dataframes together and reset the index
    master = pd.concat(to_concat, axis=0)
    master.reset_index(inplace=True)
    
    #Making sure the posts are unique with unique ID's
    duplicates = master['id'].duplicated().sum()
    
    #Final update confirming how many posts were gathered and if there are duplicates
    print(f'Final DataFrame shape: {master.shape}, there are {duplicates} duplicates')
    
    #Return the final dataframe
    return master


Posts are being added everyday, by setting a constant start time of April 19, 2020 12am, everytime the code is run, it should grab the same posts.

#Used this stackoverflow areticle to help set a permanent start time of this morning: 
#https://stackoverflow.com/questions/7241170/how-to-convert-current-date-to-epoch-timestamp
t = int(time.mktime(time.strptime('19 April, 2020', '%d %B, %Y')))

#Setting a consistent start time so everytime I run it, it should pull the same posts 
t

### 1. Gathering "Podcast' posts

pod = get_posts('podcasts', 30_000)

Checking it came through correctly:

pod.head(2)

pod.tail(2)

Exporting the raw data to a csv:

pod.to_csv('../datasets/podcasts_raw.csv', index=False)

### 2. Gathering "Television" posts

tv = get_posts('television', 30_000)

Checking it came through correctly:

tv.head(2)

tv.tail(2)

Exporting the raw data to a csv:

tv.to_csv('../datasets/tv_raw.csv', index=False)

### 3. Gathering "Video" posts

vid = get_posts('videos', 30_000)

Checking it came through correctly:

vid.head(2)

vid.tail(2)

This DataFrame was too large to be uploaded to GitHub in one csv so I split it into two exports:

vid.loc[:25000, :]

vid.loc[25000:, :]

vid.loc[25001:, :]

vid.loc[:25000, :].to_csv('../datasets/video_raw_1.csv', index=False)
vid.loc[25001:, :].to_csv('../datasets/video_raw_2.csv')
