==> code/02_Cleaning.ipynb <==
# Reddit & Quibi: Web API and NLP
## Part 2: Cleaning

In this notebook, I'm going to combine my Reddit data for all 3 subreddits and clean it.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import regex as re

import warnings
warnings.filterwarnings('ignore')

Reading in the raw data:

pod = pd.read_csv('../datasets/podcasts_raw.csv')

By looking through the columns and reviewing some of the terms on Pushshift, I'm elimninating the majority of the columns and only looking at those listed below (which include 'title' and 'selftext'). Although the features differ between the subreddits, these columns are in all of the raw csv's.

sel_col = ['id','author', 'created_utc', 'media_only', 'num_comments', 
           'over_18', 'title', 'selftext', 'spoiler', 'score', 'url']

pod_sel = pod[sel_col]

pod.tail(2)

tv = pd.read_csv('../datasets/tv_raw.csv')

tv_sel = tv[sel_col]

tv.tail()

vid_1 = pd.read_csv('../datasets/video_raw_1.csv')

vid_2 = pd.read_csv('../datasets/video_raw_2.csv')

vid = pd.concat([vid_1, vid_2], axis=0)

vid.shape

vid_sel = vid[sel_col]

Confirming my DataFrames have the same number of columns before combining them:

print(pod_sel.shape, tv_sel.shape, vid_sel.shape)

Adding a column (that will eventually be the target) for what subreddit each post belongs to:

pod_sel['subreddit'] = 'podcast'

tv_sel['subreddit'] = 'tv'

vid_sel['subreddit'] = 'video'

reddit = pd.concat([pod_sel, tv_sel, vid_sel], axis = 0)

reddit.shape

reddit.reset_index(inplace=True)

reddit.to_csv('../datasets/reddit_raw.csv', index=False)

### Missing Values
Now that we have the combined DataFrame, I'll take a look at the missing values and datatypes to see what needs to be cleaned up.

reddit.isna().sum()

Below we'll look at some of the content from columns with lots of missing data. Depending what the content is, we might be able to replace the NaN's with something more useful or we'll have to drop it.

reddit['media_only'].value_counts()

reddit['media_only'].value_counts().sum() + reddit['media_only'].isna().sum()

For the 'media_only' column, the only value we had that wasn't NaN was False. Since all of the values are either missing or False, this isn't a particularly helpful column so I will drop it.

reddit.drop(columns='media_only', inplace=True)

reddit.shape

reddit.groupby('spoiler')['subreddit'].value_counts()

It looks like each of each of the subreddits has at least a values for 'spoiler' so it's not limited to one class. After doing some research on Reddit, this is a tag that's added when the post contains spoilers. For that reason, I'm fairly confident that the missing values can be tagged as False.

reddit['spoiler'].fillna(False, inplace=True)

reddit.groupby('spoiler')['subreddit'].value_counts()

reddit['spoiler'].value_counts()

reddit['spoiler'].isna().sum()

Selftext is the actual content of the posting so this is one of the most important columns. When it comes to analyzing the text, I'm going to combine the title and the selftext into one column. If there's a title but no selftext, I'll keep the observation because it will all be one document in the end. However, if both fields are empty, that observation will have to be dropped.

reddit[reddit['selftext'].isnull()]['title'].isnull().value_counts()

Since all of the rows missing selftext have valid titles and I'm going to combine these later on, it will be ok to keep these as is.

### Datatypes
Next we'll check that the columns have the correct datatypes:

reddit.dtypes

I've also noticed that the selftext often has the character '\n' that would denote a line break. However, this is arbitrary formatting not added by the user and can cloud the actual words so I'm going to remove this. 

reddit['selftext'] = reddit['selftext'].fillna('')

reddit['selftext'] = reddit['selftext'].apply(lambda x: re.sub('\\n', ' ', x))

reddit.loc[0, 'selftext']

### Final cleaned dataset
Finally, I'll combine the title and selftext into one column, change the spoiler column to 1's and 0's, then export the cleaned data:

reddit['selftext'] = reddit['selftext'].replace('[removed]', '')

reddit['text'] = reddit['title'] + ' ' + reddit['selftext']

reddit['spoiler'] = reddit['spoiler'].astype(int)

reddit['spoiler'].value_counts()

reddit.head()

reddit_clean = reddit

reddit_clean.isna().sum()

reddit_clean.dtypes

reddit_clean.shape

This file is pretty large so I'm going to pickle it instead.

# SAVE PICKLE
import pickle
pickle.dump(reddit_clean, open('../datasets/reddit_clean.pkl', 'wb'))
