==> code/04_Modeling.ipynb <==
# Reddit & Quibi: Web API and NLP
## Part 4: Modeling

Now that I've cleaned and explored the data, I want to build the best model to classify it. To do so, I'm going to try several different classification methods:
1. [Logistic Regression](#Logistic-Regression)
2. [Naive Bayes](#Naive-Bayes)
3. [Decision Tree](#Decision-Tree)
4. [Voting Classifier](#Voting-Classifier)

import pandas as pd
import numpy as np
import pickle
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import nltk

from sklearn.model_selection import cross_val_score, train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.pipeline import Pipeline

reddit = pickle.load(open('../datasets/reddit_features.pkl', 'rb'))

reddit.shape

The goal of this model is to eventually classify the description of Quibi shows. Since those descripitions are limited in length, I'm only going to use the text for each observation. I'll use the same features and same train/test split for all models:

X = reddit['text']
y = reddit['e_subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 21)

def find_scores(model, x_tr, x_tst):
    train = model.score(x_tr, y_train)
    test = model.score(x_tst, y_test)
    cv = cross_val_score(model, x_tr, y_train).mean()
    return f'Train: {train} | Test: {test} | Cross Val {cv}'

## Logistic Regression

First, I'm going to fit logistic regression model with no stopwords. I want to identify key words for each subreddit that are so specific they give away the topic (i.e. HBO or podcasts). It's important to remove these because the Quibi descriptions won't have these, I want to focus on grabbing signal from words that talk more about the content.

Before creating the logistic regression model, I first need to vectorize the words. I'll be creating another Count Vectorizer with the stop words I determine later so this one is fairly simple.

cvec_simp = CountVectorizer(ngram_range = (1, 2), min_df = 5, max_df = .80)

cvec_simp.fit(X_train)
X_train_cvec_s = cvec_simp.transform(X_train)
X_test_cvec_s = cvec_simp.transform(X_test)

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression(solver='newton-cg')

lr.fit(X_train_cvec_s, y_train)

find_scores(lr, X_train_cvec_s, X_test_cvec_s)

### Stop Words

Based on the coefficients generated from the above model, we can see which words have the most impact. Many of these will be too specific such as the name of a TV provider or even the name of the media. While it may help in accuracy, it's not telling us anything we didn't already know. To combat this, I'm going to generate a list of stop words for each subreddit then filter it to be more robust.

lr.coef_

This DataFrame matches the terms with their coefficient for each class:

coef_df = pd.DataFrame(lr.coef_).T

features_cvec = cvec_simp.get_feature_names()

coef_df['term'] = features_cvec

These list hold the top 150 words for each subreddit:

pod_stop = coef_df.sort_values(0, ascending=False)['term'][:300]

tv_stop = coef_df.sort_values(1, ascending=False)['term'][:300]

vid_stop = coef_df.sort_values(2, ascending=False)['term'][:150]

coef_df.sort_values(1, ascending=False)['term'][:600]

Lastly, I'm going to filter the words based on how often they appear in the other subreddits. This is important because the top 150 words contain some obvious ones to remove (i.e. "podcasts" in podcast) but they also contain some key words that speak to the different content on each (i.e. "crime"). In that case, I want to keep words that appear in the other subreddits more than a certain number of times.

def filter_words(stop_list, target_class):
    sel_word = []
    for i in stop_list: 
        count = 0
        for j in list(reddit[reddit['e_subreddit'] != target_class]['text']):
            if i in j:
                count += 1
            else:
                count = count
        #print(count)
        if count < 50:
            sel_word.append(i)
    return sel_word

pod_filter = filter_words(pod_stop, 0)

tv_filter = filter_words(tv_stop, 1)

vid_filter = filter_words(vid_stop, 2)

All of these filtered lists will be my final stop words set along with a few other words I manually picked:

addtl_words = ['mic', 'mics', 'service', 'hbo', 'isp', 'dish', 'pod', 'ott', 'web series',
               'cutting', 'spectrum', 'networks', 'youtuber', 'television', 'horseman', 'renewed', 'https']

stop_words_set = (pod_filter + tv_filter + vid_filter + addtl_words)

len(stop_words_set)

stop_word_set = [x.split(' ') for x in stop_words_set]

#Advice on unpacking these lists within lists 
#https://stackoverflow.com/questions/952914/how-to-make-a-flat-list-out-of-list-of-lists
stop_word_set = [i for j in stop_word_set for i in j]

len(stop_word_set)

I want to add these to the predetermined English stop words:

english = list(CountVectorizer(stop_words = 'english').get_stop_words())

stop_word_set = stop_word_set + english

len(stop_word_set)

Now that I have set stopwords, I'll fit a few more models with a new Count Vectorizer:

cvec = CountVectorizer(tokenizer = None,
                    stop_words = stop_word_set, 
                            ngram_range = (1, 2), 
                           max_df = .95, 
                            max_features = 60_000)

cvec.fit(X_train)

X_train_cvec = cvec.transform(X_train)
X_test_cvec = cvec.transform(X_test)

lr_2 = LogisticRegression(solver='newton-cg')

lr_2.fit(X_train_cvec, y_train)

find_scores(lr_2, X_train_cvec, X_test_cvec)

This model ended up being the best score but also for interpretation, so I pickled this one to take to another notebook:

pickle.dump(lr_2, open('../model/log_reg_1.pkl', 'wb'))

pickle.dump(cvec, open('../model/cvec_1.pkl', 'wb'))

A few of the other models that I tried (but didn't make the cut for the Quibi data) are below:

tf = TfidfVectorizer(tokenizer = None, stop_words = stop_word_set, max_df = .7, ngram_range = (1, 2), min_df = 5)

tf.fit(X_train)

X_train_tf = tf.transform(X_train)

X_test_tf = tf.transform(X_test)

lr_3 = LogisticRegression(solver='newton-cg')

lr_3.fit(X_train_tf, y_train)

find_scores(lr_3, X_train_tf, X_test_tf)

## Naive Bayes

from sklearn.naive_bayes import MultinomialNB, GaussianNB

mnb = MultinomialNB()

mnb.fit(X_train_cvec, y_train)

find_scores(mnb, X_train_cvec, X_test_cvec)

This model was super fast but unfortunately not very accurate.

## Decision Tree

from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier

from sklearn.tree import DecisionTreeClassifier

dt = DecisionTreeClassifier(random_state = 42)

dt.fit(X_train_cvec, y_train)

find_scores(dt, X_train_cvec, X_test_cvec)

This model had really high variance and scored much lower than the Logistic Regreesion models above.

## Voting Classifier

from sklearn.ensemble import VotingClassifier

vc = VotingClassifier(
            estimators=[
        ('lr', lr_2), ('mnb', mnb), ('dt', dt)], voting='hard')

vc.fit(X_train_cvec, y_train)

find_scores(vc, X_train_cvec, X_test_cvec)

This model didn't do particularly great but it was interesting to see how the models interact, if one might correct the bias of another.

After trying a variety of options, the 2nd Logistic Regression model is the best and suites the needs of the project. That's what I'll use in the following notebook to predict on the Quibi shows.
